split SearchSuggestTests into core and modules/mustache also add @Nighlty to forbidden API to make sure we don't use it
